Interesting findings...
So it really looks like controlling the fan isn't all that necessary. We can keep the fan at level 1 and fairly well control the CPU temperature by simply reducing the power we deliver to the package when the CPU gets hot.

This changes the 'fan control' algorithm to do that.