Add 3 levels of noise control heuristics
In addition to manual mode we now have quiet, normal and noisy.

Quiet is equivalent to the previous heuristic. Normal an noisy try a bit harder too keep the laptop cool, which may be better for lap use.

The power control heuristic is unchanged and doesn't start ramping down the wattage until we're above 70C.